requirements.txt
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
boto3~=1.26.89
dynamorm~=0.11.0
marshmallow~=3.19.0